TMI2-683: Feature Flag Submission Export work

+++ src/main/java/gov/cabinetoffice/gap/lambda/Handler.java
@@ -9,6 +9,7 @@ import com.amazonaws.services.s3.AmazonS3ClientBuilder;
 import com.amazonaws.services.sns.AmazonSNSClient;
 import com.amazonaws.services.sns.AmazonSNSClientBuilder;
 import gov.cabinetoffice.gap.enums.GrantExportStatus;
+import gov.cabinetoffice.gap.exceptions.EmptySqsEventException;
 import gov.cabinetoffice.gap.model.GrantExportListDTO;
 import gov.cabinetoffice.gap.model.Submission;
 import gov.cabinetoffice.gap.service.*;
@@ -27,11 +28,13 @@ public class Handler implements RequestHandler<SQSEvent, SQSBatchResponse> {
     private static final AmazonS3 s3client = AmazonS3ClientBuilder.defaultClient();
     private static final OkHttpClient restClient = new OkHttpClient();
 
+    private static final boolean SUPER_ZIP_ENABLED = Boolean.parseBoolean(System.getenv("SUPER_ZIP_ENABLED"));
+
     @SneakyThrows
     @Override
     public SQSBatchResponse handleRequest(final SQSEvent event, final Context context) {
-        if (event.getRecords().size() < 1) {
-            throw new RuntimeException("No records found in SQS event");
+        if (event.getRecords().isEmpty()) {
+            throw new EmptySqsEventException("No records found in SQS event");
         }
 
         final Map<String, SQSEvent.MessageAttribute> messageAttributes = event.getRecords().get(0)
@@ -79,32 +82,35 @@ public class Handler implements RequestHandler<SQSEvent, SQSBatchResponse> {
             final Long outstandingCount = ExportRecordService.getOutstandingExportsCount(restClient, exportBatchId);
 
             if (Objects.equals(outstandingCount, 0L)) {
-                ZipService.deleteTmpDirContents();
-                logger.info("Tmp dir cleared before creating super zip");
-                try {
-                    GrantExportBatchService.updateGrantExportBatchRecordStatus(restClient, exportBatchId, GrantExportStatus.PROCESSING);
+                if (SUPER_ZIP_ENABLED) {
+                    ZipService.deleteTmpDirContents();
+                    logger.info("Tmp dir cleared before creating super zip");
+                    try {
+                        GrantExportBatchService.updateGrantExportBatchRecordStatus(restClient, exportBatchId, GrantExportStatus.PROCESSING);
 
-                    final GrantExportListDTO completedGrantExports = ExportRecordService.getCompletedExportRecordsByBatchId(restClient, exportBatchId);
-                    logger.info("Finished fetching completedGrantExports with size of: {}", completedGrantExports.getGrantExports().size());
+                        final GrantExportListDTO completedGrantExports = ExportRecordService.getCompletedExportRecordsByBatchId(restClient, exportBatchId);
+                        logger.info("Finished fetching completedGrantExports with size of: {}", completedGrantExports.getGrantExports().size());
 
-                    ZipService.createSuperZip(completedGrantExports.getGrantExports());
+                        ZipService.createSuperZip(completedGrantExports.getGrantExports());
 
-                    final String superZipFilename = HelperUtils.generateFilename(submission.getSchemeName(), "");
+                        final String superZipFilename = HelperUtils.generateFilename(submission.getSchemeName(), "");
 
-                    final String superZipObjectKey = ZipService.uploadZip(submission.getSchemeId() + "/" + exportBatchId, superZipFilename);
+                        final String superZipObjectKey = ZipService.uploadZip(submission.getSchemeId() + "/" + exportBatchId, superZipFilename);
 
-                    GrantExportBatchService.addS3ObjectKeyToGrantExportBatchRecord(restClient, exportBatchId, superZipObjectKey);
-                    GrantExportBatchService.updateGrantExportBatchRecordStatus(restClient, exportBatchId, GrantExportStatus.COMPLETE);
-
-                    NotifyService.sendConfirmationEmail(restClient, emailAddress, exportBatchId, submission.getSchemeId(),
-                            submissionId);
-                } catch (Exception e) {
-                    logger.error("Could not process message while trying to create super zip", e);
-                    GrantExportBatchService.updateGrantExportBatchRecordStatus(restClient, exportBatchId, GrantExportStatus.FAILED);
+                        GrantExportBatchService.addS3ObjectKeyToGrantExportBatchRecord(restClient, exportBatchId, superZipObjectKey);
+                        GrantExportBatchService.updateGrantExportBatchRecordStatus(restClient, exportBatchId, GrantExportStatus.COMPLETE);
+                    } catch (Exception e) {
+                        logger.error("Could not process message while trying to create super zip", e);
+                        GrantExportBatchService.updateGrantExportBatchRecordStatus(restClient, exportBatchId, GrantExportStatus.FAILED);
+                    }
                 }
+
+                // TODO move back inside the try...catch above when super zip is in production
+                NotifyService.sendConfirmationEmail(restClient, emailAddress, exportBatchId, submission.getSchemeId(),
+                        submissionId);
             } else {
                 logger.info(
-                        String.format("Outstanding exports for export batch %s: %s", exportBatchId, outstandingCount));
+                        "Outstanding exports for export batch {}: {}", exportBatchId, outstandingCount);
             }
 
             // STEP 9 - clear tmp dir as this is preserved between frequent invocations
@@ -112,17 +118,20 @@ public class Handler implements RequestHandler<SQSEvent, SQSBatchResponse> {
             logger.info("Tmp dir cleared");
         } catch (Exception e) {
             logger.error("Could not process message", e);
-            ExportRecordService.updateExportRecordStatus(restClient, exportBatchId, submissionId, GrantExportStatus.FAILED);
+
+            if (SUPER_ZIP_ENABLED) {
+                ExportRecordService.updateExportRecordStatus(restClient, exportBatchId, submissionId, GrantExportStatus.FAILED);
+            }
         } finally {
             // TODO replace existing getOutstandingExportsCount with this once feature flag is off
             logger.info("Inside finally block calling getRemainingExportsCount");
             final Long remainingExports = ExportRecordService.getRemainingExportsCount(restClient, exportBatchId);
-            logger.info(String.format("Submissions export complete. There are %s remaining exports.", remainingExports));
-            if(Objects.equals(remainingExports, 0L)) {
+            logger.info("Submissions export complete. There are {} remaining exports.", remainingExports);
+            if (Objects.equals(remainingExports, 0L)) {
                 logger.info("Calling getFailedExportsCount");
                 final Long failedSubmissionsCount = ExportRecordService.getFailedExportsCount(restClient, exportBatchId);
-                logger.info(String.format("There are %s failed submissions.", failedSubmissionsCount));
-                if(true) {
+                logger.info("There are {} failed submissions.", failedSubmissionsCount);
+                if (true) {
                     String outcome = new SnsService((AmazonSNSClient) AmazonSNSClientBuilder.defaultClient())
                             .failureInExport(schemeName, failedSubmissionsCount);
                     logger.info(outcome);
